demo_poc/requirements.txt
# Standalone install for this folder (same pins as parent ../requirements-demo.txt)
asyncua>=1.1,<2
flask>=3.0,<4
paho-mqtt>=2.0,<3
amqtt>=0.11,<1
PyYAML>=6.0,<7
matplotlib>=3.8,<4
